Smoother fixed timeline and more precise detection of current event
The timeline on the top of the screen now simply is set to fixed when in screen rather than move with the user, making it achieve the same thing, however be less rugged and jittery. The activeIndex of the current event scrolled to is more precise and now checks if the event is half visible rather than fully.